Updated __init__.py, __main__.py, words.py
This will probably be the last commit, because basically... I don't like python anymore, I have been learning the C programming language for a long time, and I have fallen in love with it. I like it much more than python. I will in these weeks (when I have time) create a C password generator, similar to pypassgen. 

In this commit there are a few tweaks, and updated the help.